Fix color test regression.
This change forces the color editor to allow duplicate scalar
values for control points (i.e., you can have color discontinuities).
Previous to this change, coincident control points were not allowed
and this was enforced by deleting one of them. However, the order
was not consistent across other changes and there is no reason to
prevent discontinuous maps, so the constraint is being removed.